+++ src/main/java/org/example/expert/config/SecurityConfig.java
@@ -2,6 +2,8 @@ package org.example.expert.config;
 
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
+import org.example.expert.global.jwt.JwtFilter;
+import org.example.expert.global.jwt.JwtUtil;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
