+++ src/main/java/org/example/expert/global/jwt/JwtFilter.java
@@ -1,4 +1,4 @@
-package org.example.expert.config;
+package org.example.expert.global.jwt;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
@@ -8,7 +8,6 @@ import io.jsonwebtoken.UnsupportedJwtException;
 import jakarta.servlet.*;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.example.expert.domain.user.enums.UserRole;
@@ -16,7 +15,6 @@ import org.example.expert.domain.user.enums.UserRole;
 import java.io.IOException;
 import org.example.expert.global.security.UserPrincipal;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.OncePerRequestFilter;
