+++ src/main/java/org/example/expert/global/jwt/JwtUtil.java
@@ -1,4 +1,4 @@
-package org.example.expert.config;
+package org.example.expert.global.jwt;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
