+++ src/main/java/org/example/expert/global/utils/PasswordEncoder.java
@@ -1,4 +1,4 @@
-package org.example.expert.config;
+package org.example.expert.global.utils;
 
 import at.favre.lib.crypto.bcrypt.BCrypt;
 import org.springframework.stereotype.Component;
